+++ lib/cartodb/middleware/error/error-middleware.js
@@ -2,16 +2,10 @@ const _ = require('underscore');
 const debug = require('debug')('windshaft:cartodb:error-middleware');
 
 module.exports = function errorMiddleware (/* options */) {
-    return function error (err, req, res, next) {
+    return function error (allErrors, req, res, next) {
         // jshint unused:false
         // jshint maxcomplexity:9
-        var allErrors = Array.isArray(err) ? err : [err];
-
-        allErrors = populateTimeoutErrors(allErrors);
-
-        const label = err.label || 'UNKNOWN';
-        err = allErrors[0] || new Error(label);
-        allErrors[0] = err;
+        var err = allErrors[0];
 
         var statusCode = findStatusCode(err);
 
@@ -19,7 +13,7 @@ module.exports = function errorMiddleware (/* options */) {
             statusCode = 204;
         }
 
-        debug('[%s ERROR] -- %d: %s, %s', label, statusCode, err, err.stack);
+        debug('[%s ERROR] -- %d: %s, %s', err.label || err.message, statusCode, err, err.stack);
 
         // If a callback was requested, force status to 200
         if (req.query && req.query.callback) {
@@ -41,37 +35,8 @@ module.exports = function errorMiddleware (/* options */) {
     };
 };
 
-function isRenderTimeoutError (err) {
-    return err.message === 'Render timed out';
-}
 
-function isDatasourceTimeoutError (err) {
-    return err.message && err.message.match(/canceling statement due to statement timeout/i);
-}
 
-function isTimeoutError (err) {
-    return isRenderTimeoutError(err) || isDatasourceTimeoutError(err);
-}
-
-function populateTimeoutErrors (errors) {
-    return errors.map(function (error) {
-        if (isRenderTimeoutError(error)) {
-            error.subtype = 'render';
-        }
-
-        if (isDatasourceTimeoutError(error)) {
-            error.subtype = 'datasource';
-        }
-
-        if (isTimeoutError(error)) {
-            error.message = 'You are over platform\'s limits. Please contact us to know more details';
-            error.type = 'limit';
-            error.http_status = 429;
-        }
-
-        return error;
-    });
-}
 
 function findStatusCode(err) {
     var statusCode;
