+++ lib/cartodb/server.js
@@ -46,7 +46,7 @@ var MapConfigAdapter = require('./models/mapconfig/adapter');
 var StatsBackend = require('./backends/stats');
 
 const lzmaMiddleware = require('./middleware/lzma');
-const errorMiddleware = require('./middleware/error-middleware');
+const errorMiddleware = require('./middleware/error');
 
 const prepareContextMiddleware = require('./middleware/context');
